Generate TLS certificates deterministically.
Generated certificates used to use random keys and serial numbers.
So they were different on each run of Redwood.
Now they use the root certificate's key, and a serial number that
is the MD5 hash of the real certificate.
So they are the same every time.
This makes it so that if you add the certificate to the browser's
list of trusted certificates, it actually sticks.