Detect and stop proxy forwarding loops.
Before, it was possible to crash Redwood just by sending it an HTTP
request asking for its own HTTP proxy port. It would keep forwarding the
request to itself until it ran out of memory or file descriptors.

Now it stops the loop by rejecting any request that already has 10 or
more X-Forwarded-For headers.

Change-Id: Id92cfa56ed51e58bef64efb67b63574881908891