Allow triggering reload over HTTP.
The normal way to trigger a reload is with SIGHUP. But now you can also
do it by accessing http://localhost:6502/reload (or whatever port it
actually is). But it must be localhost, to avoid catching other URLs
that may have a path of /reload.

This should be useful on Windows, which doesn't have SIGHUP, and on
systems where the administrator is not very comfortable on the command
line.