Move proxy.pac to API server, and simplify.
Only enable per-user ports when they are specified in the password file.
The ephemeral ones made too many problems when Redwood was restarted.

Use the address from the Host header as the proxy address in the PAC
file. That should guarantee that it is an address that the client can
actually reach us on. It also eliminates two configuration settings.

Forget about wpad.dat. That seldom worked anyway.