Don't let URL regexes contain spaces.
The old rule is that a regex extends to the last slash on the line
(except in comments).  This was causing problems with content-pruning
rules I wanted to write:

/craigslist/d p.row:has(a[href*="/tix/"])

Was being interpreted as a regex of
/craigslist/d p.row:has(a[href*="/tix/
and a CSS selector of
"])
which obviously doesn't work.

Now it only looks for slashes until it comes to a space.