When generating a certificate, just match the SNI.
Some HTTPS servers have certificates that are valid for more than one
site. Formerly, when Redwood SSLBumped a connection to such a server, it
generated a certificate that covered the same sites as the original.
This could cause problems because some browsers (notably Chrome) will
reuse that connection for all the servers that the certificate covers.
This behavior is undesirable when we want to SSLBump some of those
servers and not others.

Now, if the client hello contains a Server Name Indication (SNI),
Redwood generates a certificate that covers only that one name. This
doesn't solve the problem completely, since we have no control over what
certificate is used on connections that we don't SSLBump, but it helps
make SSLBump behavior more consistent.