Replace auth-helper with simpler authenticator interface.
Redwood has had the auth-helper flag for delegating authentication to an
external program for a long time. But the looping, piping interface we
borrowed from Squid is rather complicated. So we're replacing it.

The new authenticator just takes the username and password as
command-line arguments. If the command "succeeds" (has a zero exit
value), the user is authenticated. Here is an example of a shell script
that implements the interface (and accepts a username of "test" with a
password of "test"):

    #!/bin/sh
    [ "$1:$2" = "test:test" ]

I would have kept the old auth-helper interface around for
compatibility, but I noticed that it leaks processes when the
configuration is reloaded. (Each time the configuration is reloaded, it
starts a new instance of the helper, but it doesn't shut down the old
one.) So I just took it out.