Use TLS SNI when fetching certificate to imitate.
Redwood currently uses the server address from the CONNECT request,
or the original IP address from an intercepted connection,
to connect to the origin server and imitate its certificate.
Switch to using the SNI server name instead, when available.

Also, disable support for HTTPS on the internal server (203.0.113.1),
since it doesn't seem to work anyway.