Remove max-conns-per-host.
There are bugs in the implementation of http.Transport.MaxConnsPerHost.
They are causing panics. So we'll stop using that setting, and deal with
the problem it was supposed to fix another way: If the upstream server
doesn't support HTTP/2, we'll limit the client to HTTP/1.1. That way the
client will limit the nubmer of concurrent requests itself.