Don't store config; call getConfig more often.
Redwood's memory usage is high, due to how long it takes to garbage-collect
old configurations that are left after reloading. So this commit reduces how
long references to the configuration are kept. This means that if the configuration
is reloaded while a requrest is being filtered, part of the filtering will be done
with the old configuration, and part with the new one.