More efficient websockets.
When making a WSS (websocket over TLS) connection, Redwood used to discard
the original connection to the origin server and dial a new connection.
Now it reuses the original connection. This is more efficient, and it also
makes connection changes from Starlark scripts apply to websockets as well
as HTTPS connections.

A large part of this change is to ensure that Redwood doesn't make an HTTP/2
connection to the origin server if the client doesn't support HTTP/2.
(These days, the most likely reason for a client to omit support for HTTP/2
is that it's making a websocket connection, which must be HTTP/1.1.)