Revert 'Block non-TLS CONNECT requests'.
This was blocking too much legitimate traffic. But it's a good idea
in general. So I added a new virtual ACL ("transparent") that lets
you do the same thing, or a finer-grained version. Here is an ACL
action line that is equivalent to what we had before this change:

block invalid-ssl !transparent