Copy some TLS settings from the client.
Rather than hard-coding settings copied from Chrome to use for the
connection to the upstream server, we now copy the list of cipher suites
and the list of elliptic curves and key exchange mechanisms from the
incoming client hello message.

We add X25519MLKEM768 to the list of key exchange mechanisms if it is
not present, because that improves the CloudFlare score for older
browsers connecting through Redwood.