Add alternate HTML tree for Starlark.
The bsoup package in starlib is read-only, and we want to be able to
modify HTML. So we need an HTML tree that we can modify.
Since we need to build this from scratch, we might as well use the API
that we prefer. So the new one has an API based on selectolax—but only
three methods and one attribute so far…