Pull in security fixes.
There were a bunch of security fixes announced today in the standard
library and the x/net packages. The one in x/net/html was likely
my doing, many years ago, but the way the package has been moved around
over the years makes it hard to track down for sure.